BREAKING CHANGE: change PollRate type to int64 (#153)
Align poll rate handling with Splunk Observability API. As of Mar 2022 the API supports any value between 1 and 10 minutes (in milliseconds).

Previously AWS was restricted to 1-5 mins. Azure and GCP used to support either 1 or 5 minute intervals.

This is a BREAKING CHANGE as the PollRate type was previously used to define poll rate value. Now it's just a regular int64 (value in milliseconds).