Fix sendAlertsOnceMutingPeriodHasEnded JSON marshaling bug
The AlertMutingRule SendAlertsOnceMutingPeriodHasEnded currently has the
"omitempty" JSON tag. Because false is the boolean zero value, setting
SendAlertsOnceMutingPeriodHasEnded to false causes the field to be
omitted when marshaled as JSON. Since the API defaults this value to
true, it is not possible to set it to false using the Create or
UpdateAlertMutingRule functions.

Removing the "omitempty" tag will allow users to set this field to
false.