SignalFlow Client v2 Rework
This is a major work of the SignalFlow client. It involves several breaking
changes (documented in the new README.md file in this commit) and is going to
be released as a separate package and as a v2.

The main theme of these changes is better and cleaner shutdown of
computations/clients and timeout handling of metadata. Context objects are no
longer stored in structs (as recommended by Go authorities) but are passed in
through the public APIs and can be used to timeout/cancel potentially blocking
operations such as sending messages. Shutdown is less dependent on context
cancelation and uses channel closing to signal goroutines to stop. Generally
only the types that send to goroutines will close a channel and the receivers
will watch for the closure and return.

This also adds golangci-lint and fixes a bunch of preexisting errors with that.

Fixes #190.